Completing Notification Logic
Now a user ubscribes to notifications in a more controlled frontend enviornment.
The server tracks subscriptions and associates them with each user and can use that to its advantage.

The next step will be to manage the cron jobs that the server implements to reduce server load
and increase scalability.
