+++ src/context/InteractiveContext.jsx
@@ -1,5 +1,5 @@
 import { createContext, useState, useEffect } from "react";
-import { getNotifications, getNotificationsAtStart } from "../utils/api";
+// import { getNotifications, getNotificationsAtStart } from "../utils/api";
 
 const InteractiveContext = createContext({});
 
@@ -8,57 +8,57 @@ export const InteractiveProvider = ({ children }) => {
   const [showLogin, setShowLogin] = useState(false);
   const [showNotifs, setShowNotifs] = useState(false);
   const [confirm, setConfirm] = useState(false);
-  const [notifications, setNotifications] = useState([]);
+  // const [notifications, setNotifications] = useState([]);
   const [view, setView] = useState("month");
   const [addNewEvent, setAddNewEvent] = useState(false);
   const [type, setType] = useState(null);
   const [listUpdate, setListUpdate] = useState([]);
   const [event, setEvent] = useState(null);
 
-  useEffect(() => {
-    send();
-  }, []);
+  // useEffect(() => {
+  //   send();
+  // }, []);
 
-  const send = async () => {
-    try {
-      const token = localStorage.getItem("authToken");
-      const user = localStorage.getItem("user");
-      if (token && user) {
-        const parsedUser = JSON.parse(user);
-        if (parsedUser.notifSub) {
-          const userId = parsedUser.id;
-          const serverSentSource = getNotifications(userId);
-          getNotificationsAtStart(parsedUser.username, token)
-            .then((res) => {
-              const oldNotifs = res.data.notifs;
-              setNotifications(oldNotifs);
-              setupNotifListener(serverSentSource);
-            })
-            .catch((err) => {
-              console.log(err);
-            });
-        }
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
+  // const send = async () => {
+  //   try {
+  //     const token = localStorage.getItem("authToken");
+  //     const user = localStorage.getItem("user");
+  //     if (token && user) {
+  //       const parsedUser = JSON.parse(user);
+  //       if (parsedUser.notifSub) {
+  //         const userId = parsedUser.id;
+  //         const serverSentSource = getNotifications(userId);
+  //         getNotificationsAtStart(parsedUser.username, token)
+  //           .then((res) => {
+  //             const oldNotifs = res.data.notifs;
+  //             setNotifications(oldNotifs);
+  //             setupNotifListener(serverSentSource);
+  //           })
+  //           .catch((err) => {
+  //             console.log(err);
+  //           });
+  //       }
+  //     }
+  //   } catch (err) {
+  //     console.log(err);
+  //   }
+  // };
 
-  const setupNotifListener = (serverSentSource) => {
-    console.log("Set up and listening for notifications");
-    serverSentSource.addEventListener("open", () => {
-      console.log("Open");
-    });
-    serverSentSource.addEventListener("message", (event) => {
-      const notification = JSON.parse(event.data);
-      setNotifications((prev) => [notification, ...prev]);
-      console.log("Received notification:", notification);
-    });
-    serverSentSource.addEventListener("error", (error) => {
-      console.error("SSE error:", error);
-      serverSentSource.close();
-    });
-  };
+  // const setupNotifListener = (serverSentSource) => {
+  //   console.log("Set up and listening for notifications");
+  //   serverSentSource.addEventListener("open", () => {
+  //     console.log("Open");
+  //   });
+  //   serverSentSource.addEventListener("message", (event) => {
+  //     const notification = JSON.parse(event.data);
+  //     setNotifications((prev) => [notification, ...prev]);
+  //     console.log("Received notification:", notification);
+  //   });
+  //   serverSentSource.addEventListener("error", (error) => {
+  //     console.error("SSE error:", error);
+  //     serverSentSource.close();
+  //   });
+  // };
 
   return (
     <InteractiveContext.Provider
@@ -66,7 +66,7 @@ export const InteractiveProvider = ({ children }) => {
         menu,
         showLogin,
         confirm,
-        notifications,
+        // notifications,
         view,
         addNewEvent,
         type,
@@ -81,7 +81,7 @@ export const InteractiveProvider = ({ children }) => {
         setView,
         setConfirm,
         setMenu,
-        setNotifications,
+        // setNotifications,
         setShowLogin,
       }}
     >
